pyproject.toml
[project]
name = "oregon-trail-decompiler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.1.0",
    "pytest>=8.3.5",
]
